refactor: extract AdminModal from admin index
Move the modal rendering (all create/edit forms for user accounts,
staff, vendors, products, and SMS templates) into a dedicated
component that receives form state and callbacks as props.
